pyproject.toml
[project]
name = "task-logger"
version = "0.1.0"
description = "Local task/activity logger with system tray and global hotkey"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "pynput>=1.7.0",
    "pystray>=0.19.0",
    "Pillow>=11.0.0",
]

[project.scripts]
task-logger = "launcher:main"
